FE-1115 - Extract new component
