+++ src/pages/signals/dashboards/components/HealthScoreChart/HealthScoreChart.js
@@ -3,13 +3,11 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import moment from 'moment';
 import _ from 'lodash';
-import { tokens } from '@sparkpost/design-tokens-hibana';
-import { Bar, Line, ReferenceLine, Tooltip, XAxis, YAxis } from 'recharts';
 import { getCaretProps, getDoD } from 'src/helpers/signals';
 import { selectCurrentHealthScoreDashboard } from 'src/selectors/signals';
 import { Panel } from 'src/components/matchbox';
 import BarChart from '../../../components/charts/barchart/BarChart';
-import { LineChart, lineChartConfig } from 'src/components/charts';
+import HealthScoreLineChart from '../../../components/HealthScoreLineChart';
 import TooltipMetric from 'src/components/charts/TooltipMetric';
 import { PanelLoading } from 'src/components';
 import Callout from 'src/components/callout';
@@ -105,24 +103,6 @@ export function HealthScoreChart(props) {
     return { value: _.isNil(max) ? 'n/a' : max };
   }
 
-  function formatTooltipPayload(data) {
-    const firstItem = data[0];
-    const payload = firstItem ? firstItem.payload : {};
-
-    if (payload) {
-      return [
-        ...data.map(item => {
-          return {
-            ...item,
-            stroke: thresholds[payload.ranking].color,
-          };
-        }),
-      ];
-    }
-
-    return payload;
-  }
-
   return (
     <Panel sectioned title={`${formatDate(filters.from)} – ${formatDate(filters.to)}`}>
       <div className={styles.Content}>
@@ -131,82 +111,11 @@ export function HealthScoreChart(props) {
         {!noData && (
           <>
             {isHibanaEnabled ? (
-              <div className="LiftTooltip" style={{ position: 'relative' }}>
-                <LineChart>
-                  <LineChart.Container height={300} data={history}>
-                    <Bar
-                      key="this-is-a-key"
-                      dataKey="noKey"
-                      stackId="stack"
-                      cursor="pointer"
-                      isAnimationActive={false}
-                      background={lineChartConfig.barsBackground}
-                      onMouseOver={handleDateHover}
-                    />
-
-                    <XAxis
-                      dataKey="date"
-                      axisLine={false}
-                      scale="auto"
-                      height={30}
-                      tickLine={false}
-                      {...getXAxisProps()}
-                    />
-
-                    <YAxis
-                      dataKey="health_score"
-                      axisLine={false}
-                      tickLine={false}
-                      width={30}
-                      minTickGap={2}
-                      ticks={[0, 55, 80, 100]}
-                    />
-
-                    <Tooltip
-                      cursor={<LineChart.Cursor data={history} />}
-                      content={({ payload, ...props }) => {
-                        return (
-                          <LineChart.CustomTooltip
-                            {...props}
-                            payload={formatTooltipPayload(payload)}
-                          />
-                        );
-                      }}
-                      wrapperStyle={lineChartConfig.tooltipStyles}
-                      isAnimationActive={false}
-                      labelFormatter={formatDate}
-                      nameFormatter={() => 'Health Score'}
-                      formatter={val => val}
-                    />
-
-                    <ReferenceLine
-                      {...lineChartConfig.referenceLineProps}
-                      strokeDasharray="none"
-                      y={100}
-                    />
-
-                    <ReferenceLine
-                      {...lineChartConfig.referenceLineProps}
-                      y={80}
-                      style={{ stroke: tokens.color_green_700 }}
-                    />
-
-                    <ReferenceLine
-                      {...lineChartConfig.referenceLineProps}
-                      y={55}
-                      style={{ stroke: tokens.color_red_700 }}
-                    />
-
-                    <Line
-                      {...lineChartConfig.lineProps}
-                      dataKey="health_score"
-                      stroke={tokens.color_blue_700}
-                    />
-                  </LineChart.Container>
-
-                  <LineChart.YAxisLabel></LineChart.YAxisLabel>
-                </LineChart>
-              </div>
+              <HealthScoreLineChart
+                data={history}
+                filters={filters}
+                onBarMouseOver={handleDateHover}
+              />
             ) : (
               <BarChart
                 margin={newModelMarginsHealthScore}
